viewer: set term size environment vars before filter command starts
When the part viewer embedded terminal starts we take the opportunity to
initialize the COLUMNS and LINES environment variables of the filter
command before starting the terminal.

Unfortunately, the part viewer Draw() method will start the terminal
after fetching the body part which will (when complete) start the filter
command in attemptCopy(). There is a race which can lead to the env
variables being updated after the filter command has been started.

Set the environment variables before triggering the part fetch. This
will guarantee that they are present when the filter command starts.

Signed-off-by: Robin Jarry <robin@jarry.cc>
Tested-by: Inwit <inwit@sindominio.net>