compose: prevent trailing spaces on empty headers
When using format-flowed=true, you need to configure your editor to
format the text accordingly. With (neo)vim, this is achieved by setting
formatoptions+=w (trailing whitespace indicates paragraph continuation).
This may also be combined with +=a (automatically reflow paragraphs as
you edit).

When also using edit-headers=true, empty header lines are considered
part of the same paragraph (by vim) if they have a trailing space. This
can make editing the headers difficult (the editor may mistakenly join
them together).

To prevent this, vary the separator between header field name and body
depending on whether the body is empty. If so, only use the ":".  This
is valid syntax for the populated case too, but including a space is
aesthetically pleasing, so for that case use ": " as the separator.

Signed-off-by: Jonathan Dowland <jon+aerc-devel@dow.land>
Acked-by: Robin Jarry <robin@jarry.cc>