send: avoid leaking bcc addresses to all recipients
Depending on the provider/backend which is sending the email, the Bcc
headers may or may not be stripped. Make sure to remove them from aerc
before sending (and optionally copying) the message.

Add an accounts.conf option to preserve them if needed.

Fixes: https://todo.sr.ht/~rjarry/aerc/283
Changelog-fixed: Some SMTP servers do not strip `Bcc` headers. aerc
 now removes them before sending emails to avoid leaking private
 information. A new `strip-bcc = false` option can be used in
 `accounts.conf` to revert to previous behaviour (preserve _Bcc_
 headers in outgoing messages).
Signed-off-by: Robin Jarry <robin@jarry.cc>
Reviewed-by: Diederik de Haas <didi.debian@cknow.org>