reply: add -f flag to always include From header addresses
A badly implemented mailing list sets itself as Reply-To, which means
that the original sender currently gets dropped.

Add -f flag that sends reply to both From and Reply-to sets of
addresses.  Sending to both is allowed in the mail RFCs, so this flag
can be safely used for all mail.

Changelog-added: Reply to addresses in `From` and `Reply-To` headers
 with `:reply -f`.
Signed-off-by: Radim Krčmář <radim@krcmar.dev>
Acked-by: Robin Jarry <robin@jarry.cc>