compose: disable {prev,next}-field on review screen
It is currently possible to focus the various header editors while on
the compose::review screen but it is impossible to actually change their
value.

Prevent the focus from even occurring and return an explicit error
message to make it explicit that :next-field and :prev-field only work
when the text editor is visible and [compose].edit-headers = false.

Also, to avoid any ambiguity, prevent any text input from being focused
unless the text editor is visible. This effectively prevents the cursor
from being displayed.

Signed-off-by: Robin Jarry <robin@jarry.cc>
Reviewed-by: Bence Ferdinandy <bence@ferdinandy.com>