compose: fix deadlock when editor exits with an error
When exiting vim with :cq, it exits with an error status which is caught
in the termClosed() callback. This causes the composer tab to be closed
and it is a known and expected behaviour.

A recent fix of the tab strip removal revealed that the composer lock is
still held when calling RemoveTab(). Internally, RemoveTab() hides and
unfocuses the removed tab both operations require the composer lock as
well, causing a complete deadlock of the application.

Unfortunately, we cannot rely on the go defer statement to have
RemoveTab(), called *AFTER* c.Unlock(). Deferred statements are called
in the reverse order they were evaluated.

Use an explicit function array to ensure that the composer is unlocked
before RemoveTab() is called.

Fixes: 6b97085ae5ec ("tab: fix broken history on removal")
Signed-off-by: Robin Jarry <robin@jarry.cc>
Tested-by: Julio B <julio.bacel@gmail.com>