pgp: ensure mime-version header is in the right place
Update go-pgpmail which has a similar fix to the one we had in place.
Update our fix to use the same implementation: do not parse the whole
signed part. Only parse the part header, remove the mime-version header
from it if any and re-write it, followed the part body verbatim without
any decoding/re-encoding.

Fixes: 7346d20343b9 ("gpg: fix signed message encoding")
Fixes: 7b8fbe2a3d7c ("gpg: fix mime-version header position")
Signed-off-by: Robin Jarry <robin@jarry.cc>
Reviewed-by: Simon Ser <contact@emersion.fr>