ui: fix the position of the address-book completion popovers
When composing a message, the location of the popovers was not offset by
the location of the field that was being completed (To, Cc,...)
resulting in them appearing always at the same position near the top
left corner of the window. Restore the offset to have them appear right
underneath the active field again.

Fixes: 787cfbd9a90a ("ui: remove screen and viewports")
Changelog-fixed: The address-book completion popovers now again appear
 under the field being completed.
Signed-off-by: Karel Balej <balejk@matfyz.cz>
Reviewed-by: Tim Culverhouse <tim@timculverhouse.com>
Acked-by: Robin Jarry <robin@jarry.cc>