compose: merge repeated address headers
When composing a message with edit-headers=true and some To, Cc, Bcc
headers are repeated, only the last occurrence will be considered. All
other addresses will be dropped and lost without any error/warning.

When parsing the embedded headers, merge any repeated To, Cc, Bcc
headers and warn the user about it.

Reported-by: Inwit <inwit@sindominio.net>
Signed-off-by: Robin Jarry <robin@jarry.cc>
Tested-by: Inwit <inwit@sindominio.net>