compose: only show default annotations if absent from config
Currently, it is impossible to "hide" bindings presented on the review
screen even if removing the annotations from binds.conf.

If the configuration contains any annotated bindings, completely ignore
all of the default annotations.

Only display bindings from the config that are both present and
annotated.

Fixes: 64502d9e8b04 ("compose: show annotations on the review screen")
Changelog-changed: Key bindings available on the compose review screen
 are now only displayed if they are annotated with a comment. If
 no binding is annotated in `[compose::review]`, then the default
 annotations will be used as a fallback to avoid displaying a blank
 screen.
Reported-by: Skejg <grolleman@zoho.com>
Signed-off-by: Robin Jarry <robin@jarry.cc>
Tested-by: Inwit <inwit@sindominio.net>