deps: update go-jmap to v0.5.1
Update go-jmap to v0.5.1. This patch version includes race condition
fixes for accessing a Session object from the various Client methods.
More important for aerc, it enforces that all client requests include
the JMAP Core URI in the "using" capability string. This was corrected
via an errata to RFC 8620 and 8621.

Link: https://www.rfc-editor.org/errata/eid6606
Reported-by: Benoit Tellier <btellier@linagora.com>
Reported-by: Robin Jarry <rjarry@jarry.cc>
Signed-off-by: Tim Culverhouse <tim@timculverhouse.com>
Acked-by: Robin Jarry <robin@jarry.cc>