deps: update go-jmap to v0.5.2
Update go-jmap to v0.5.2. This patch version includes race condition
fixes for accessing a Session object from the various Client methods.
More important for aerc, it enforces that all client requests include
the JMAP Core URI in the "using" capability string. This was corrected
via an errata to RFC 8620 and 8621.

Also correct a bug from v0.5.1 where we checked for required
capabilities only in the parsed map. go-jmap parses the core capability
into a different field, which causes an error when requesting a core
capability or any unparsed capability. Go-jmap now checks the raw JMAP
map received from the server which enables clients to use any capability
the server has advertised it supports.

Link: https://www.rfc-editor.org/errata/eid6606
Reported-by: Benoit Tellier <btellier@linagora.com>
Reported-by: Robin Jarry <rjarry@jarry.cc>
Signed-off-by: Tim Culverhouse <tim@timculverhouse.com>
Acked-by: Robin Jarry <robin@jarry.cc>