viewer: disable passthrough mode on pager exit
To avoid accidentally closing the pager while the passthrough mode is
enabled (and therefore all key presses go to /dev/null), make sure to
exit the passthrough mode when the pager exits.

Also, to make the user experience more pleasant, exit the passthrough
mode upon pressing the <enter> key in the pager virtual terminal. This
should match the workflow where the users press "/" to start searching
in the pager, and press enter to validate their search.

Move the toggling of passthrough mode in a global function. Since that
setting is global to all accounts, make sure to update all the status
lines. Reuse that function for the :toggle-key-passthrough command and
when exiting the message viewer or pressing enter.

Reported-by: Jonathan Dowland <jon@dow.land>
Signed-off-by: Robin Jarry <robin@jarry.cc>