imap: properly set mailbox information in list-status calls
aerc is sluggish at start-up and when switching directories when
accessing my 5GB large WorkMail mailbox using IMAP.

This is due to to LIST-STATUS actions taking 7 (!!) seconds, because it
does not properly set the reference name in LIST-STATUS calls, hence
lists *everything* in the mailbox. When a specific folder is selected
(e.g. pretty much all the time except at startup), we can and should
restrict to the current folder. This is exactly what this patch does.

You can see before and after numbers for my mailbox in
  https://paste.sr.ht/~simartin/86065a15ded912375161f96031f4cf779c7e65d4

In terms of user experience, that latency observed between starting aerc
on my mailbox and having the default folder displayed goes from 29s to
14s (it's still high but I think it's still a good improvement).
Switching between folders also feels much snappier.

Signed-off-by: Simon Martin <simon@nasilyan.com>
Tested-by: Folker Schwesinger <dev@folker-schwesinger.de>
Tested-by: Moritz Poldrack <moritz@poldrack.dev>
Acked-by: Robin Jarry <robin@jarry.cc>