colorize: use snprintf instead of sprintf
Define an inline XSPRINTF macro to use the same naming scheme than
xmalloc. Check the return value of snprintf(). If it fails, print an
error and abort the program.

Replace all calls to sprintf() by XSPRINTF().

This is probably more to make C compilers happy and for consistency than
anything.

Signed-off-by: Matěj Cepl <mcepl@cepl.eu>
Signed-off-by: Robin Jarry <robin@jarry.cc>