app: update folder counters upon message deletion
I noticed that when deleting an unread email from the current folder (or
moving it to another folder), the "visual unread count" is not
decremented (and later on I'm desperately looking for the mysterious
unread mail that does not exist anymore :-)).

This patch adds a parameter to AccountView.updateDirCounts stating
whether the caller reacts to a deletion, and makes sure to call it when
processing a MessagesDeleted message.

Signed-off-by: Simon Martin <simon@nasilyan.com>
Tested-by: Inwit <inwit@sindominio.net>
Acked-by: Robin Jarry <robin@jarry.cc>