ui: map number block to regular counterparts
Due to vaxis low-level design, the enter key of the number block has
a different key code that a regular carriage return. Since most people
do not care which of the two they are hitting, remap the former and
up/down/left/right to act like the latter.

Changelog-added: Support for number block enter to act as <Cr>
 in mappings.
Signed-off-by: Moritz Poldrack <git@moritz.sh>
Reviewed-by: Tim Culverhouse <tim@timculverhouse.com>
Acked-by: Robin Jarry <robin@jarry.cc>