ui: allow multiple stylesets to be specified in styleset-name
The patch allows to highlight messages in specific accounts without
affecting the others.

Without the patch you need to copy the whole styleset, add the
highlighting commands, then use that styleset with that specific
account.

With the patch you can make a small styleset with just the highlighting
stuff and chain it up to the main styleset this way:

        styleset-name: main
        [ui:account=special]
        styleset-name: main,highlight

Changelog-added: Stylesets can be chained-up in `styleset-name`
 config parameter using comma-separated format.
Signed-off-by: Vitaly Ovchinnikov <v@ovch.ru>
Acked-by: Robin Jarry <robin@jarry.cc>