carddav-query: fixed parsing addresses from response
There is an issue with the carddav-query script that was causing it to
fail when trying to query Fastmail. The script made a successful query,
but did not return any results on STDOUT. Some of the entries returned
by Fastmail are in the following format:

email0.EMAIL;TYPE=HOME:name@example.com

However, the regex for parsing the response is only set to handle
`EMAIL` or `ITEMx.EMAIL` (where x is a number). I changed the regex to
be more permissive and match `any_word.EMAIL`, which fixed the problem.

This solution allows Fastmail CardDAV to be queries and parsed for
address completion, while still excluding unrelated (non-email address)
fields.

Signed-off-by: Daniel Fichtinger <daniel@ficd.ca>
Acked-by: Robin Jarry <robin@jarry.cc>