imap: properly update UI upon message (un)read in another client
I access my personal mailbox from at least 3 different clients, and have
noticed that aerc only half handles the read/unread actions made from
another client on messages in the currently selected directory. While
the message list is properly updated, the unread count is not, and I end
up again looking for unread messages that don't exist :-)

The problem is that in such scenarii, aerc receives a MessageUpdate
message from the IMAP server, and if it already has a message with the
same UID in its message store, happily replaces it... without looking at
the Seen status, that is materialized outside of the message table.

This patch detects when the Seen status changes, and updates the UI
accordingly.

Signed-off-by: Simon Martin <simon@nasilyan.com>
Acked-by: Robin Jarry <robin@jarry.cc>