msgstore: properly rebuild message store upon reconnection
I noticed another status synchronization issue between aerc and the IMAP
server when multiple clients are involved *and* the computer running
aerc hibernates. I acknowledge it's probably not very frequent but
that's my day to day ;-)

The scenario is as follows:
- aerc sees new unread emails, and renders them properly
- The computer running aerc goes to sleep
- Those emails get read via another client
- The computer running aerc wakes up, and aerc reconnects to the IMAP
  server and creates a new session.
- aerc does NOT redraw the message list (because no other mail was
  received while it was asleep) and from there on
   - The unread messages remain bold
   - Explicitly marking them read does not do anything because this
     involves a call to the server that already knows the message is
     read, and does not send back a MessageUpdate (that aerc usually
     reacts to to "unbold" the message)

This patch fixes this issue by making sure that we clear the directory
list's message store upon disconnection from the backend, and
recomputing the RUE counts from memory when processing a
types.DirectoryContents, via new AccountView.refreshDirCounts function.
It also leverages this function in the fix I made earlier via
https://lists.sr.ht/~rjarry/aerc-devel/patches/58306, and protects from
showing negative counters in the UI, just in case.

This function is pretty cheap since it's working on memory (we're well
below the millisecond on my largish mailbox - see
https://paste.sr.ht/~simartin/124a5fa4568a4ed8cb6e639c62bb01b2d6803588),
and given the number of synchronization issues I'm meeting with those
counters, my aim is to leverage it more in the future, and ideally get
rid of the RUE counts in models.Directory, and the dance we're doing to
try to keep them up-to-date.

Signed-off-by: Simon Martin <simon@nasilyan.com>
Tested-by: Julio B <julio.bacel@gmail.com>
Acked-by: Robin Jarry <robin@jarry.cc>