commands: add -u option to {next,prev}-folder
Add the "-u" option to the :next-folder and :prev-folder commands, used
to jump to the next or the previous folder that contains at least one
unseen email.

For directory trees, we have to factor in that the number of currently
visible folders may be lower than the total number of folders, so count
that in visibleDirsCount().

I find the following bindings useful:

  <A-j> = :next-folder -u<Enter>
  <A-k> = :prev-folder -u<Enter>

It allows jumping between folders efficiently while keeping the ability
to jump to the next or previous folder normally.

References: https://todo.sr.ht/~rjarry/aerc/168
Changelog-added: Add the -u option to :next-folder and :prev-folder
 commands to cycle between folders that contain unseen messages.
Signed-off-by: Antonin Godard <antonin@godard.cc>
Acked-by: Robin Jarry <robin@jarry.cc>