msg: fix issue with reply-to-self introduced by 063c19b6
Commit 063c19b6 fixes an issue with reply-all when reply-to-self is set
to try, but unfortunately introduces a new one, reported by Drew on IRC:
it will systematically Cc the replier if they reply to an email they
sent to themselves.

The problem with that fix is that it removes the from from the deduper
loo late, and should have instead not put it there in the first place if
reply-to-self is true; this is exactly what this patch does.

Fixes: 063c19b6 ("honor reply-to-self when replying to all")
Fixes: https://todo.sr.ht/~rjarry/aerc/311
Reported-by: Drew Devault <sir@cmpwn.com>
Signed-off-by: Simon Martin <simon@nasilyan.com>
Acked-by: Robin Jarry <robin@jarry.cc>