account: allow globs in folders-sort
In a reasonably folder-heavy workflow, folders-sort quickly becomes
quite unwieldy. This change allows users to sort using a pattern syntax,
where folders that all match the pattern are sorted alphabetically.

Given the folders:
- `a`
- `b`
- `c/a`
- `c/b`

and the configuration:

```ini
folders-sort = b,c/*,a
```

one will get the following output in aerc:

- `b`
- `c/a`
- `c/b`
- `a`

This is an RFC because it needs docs changes and the like; I just wanted
to see if people liked it before going ahead with that.

Changelog-changed: The `<account>.folders-sort` now supports patterns
 for folder names.
Signed-off-by: Nicole Patricia Mazzuca <nicole@streganil.no>
Acked-by: Robin Jarry <robin@jarry.cc>