html: only pass supported w3m options
Viewing message with only an HTML part currently fails on my Mac,
spitting out the output of "w3m -o" to my face.

This happens because the version of w3m that I pulled from Homebrew was
not built with support for Gopher, hence errors out when passed the
gopher_proxy option... which the "html" script does unconditionally.

This patches fixes the script to not pass "*_proxy" options that are not
built into the w3m binary.

Signed-off-by: Simon Martin <simon@nasilyan.com>
Acked-by: Robin Jarry <robin@jarry.cc>