app: robustify unread counter updates upon message deletion
I recently added a call to updateDirCounts for deletions, and that
function does not check the account store against nil before
dereferencing it. It could be nil for example in case of a disconnection
with the IMAP server during a large deletion, and aerc crashes.

This makes updateDirCounts properly handle cases where the store is nil.

Fixes: b099981a ("update folder counters upon message deletion")
Reported-by: Andreas Radke <andreas.radke@mailbox.org>
Signed-off-by: Simon Martin <simon@nasilyan.com>
Acked-by: Robin Jarry <robin@jarry.cc>