imap: properly clean-up resources upon user initiated disconnect
While investigating how to speed up IMAP sync with the ProtonMail
bridge, and using :[dis]connect quite a bit, I noticed that after a
:disconnect, the first :connect consistently fails with the following
error (the next one works fine):

   auto-reconnect is disabled; run connect to enable it

This happens because the IMAP worker only partially cleans up its
connection dependent context, and what is left behind comes in the way
of the next :connect.

This patch makes the clean-up complete, and make :disconnect => :connect
work as expected.

Signed-off-by: Simon Martin <simon@nasilyan.com>
Acked-by: Robin Jarry <robin@jarry.cc>