menu: do not show warning when no command is provided
When users do something like ":menu -ad :cf -a", aerc is currently
showing a warning in the status line because no command was provided.
While it makes sense to warn the user if the command they provided is
not available, I don't think it does when they have explicitly decided
not to provide one (and are happy with the fallback).

Signed-off-by: Simon Martin <simon@nasilyan.com>
Acked-by: Robin Jarry <robin@jarry.cc>